backend/requirements.txt
# Django
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
django-environ==0.11.2
django-filter==23.4

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Celery
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.1
amqp==5.2.0

# Authentication
djangorestframework-simplejwt==5.3.0
django-allauth==0.57.0
dj-rest-auth==5.0.2

# Utilities
python-decouple==3.8
Pillow==10.1.0
whitenoise==6.6.0
gunicorn==21.2.0

# Markdown support
markdown==3.5.1
bleach==6.1.0

# Development
django-extensions==3.2.3
django-debug-toolbar==4.2.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0
faker==20.1.0